requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt
pillow==10.3.0
pycparser==2.22
PyJWT==1.7.1
sqlparse==0.5.0
